Refactor argument parsing to have global options and profiler subparser
All the profiler argparse options are now handled inside of a subparser.
Also, updated the argparse.ArgumentParser implementation to allow
setting a default subparser, which is set to the profiler subparser.
This enables the use of subparser options without having to specify the
subparser name in the command line arguments.

Test: m torq
Test: Run all torq tests
Bug: 395894733
Change-Id: I681ec2ef8c89e590e825d1c1eaf74e4615c6efd6